pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "esper-morphogen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2.0,<3.0",
    "torchvision>=0.17.0",
    "torch-geometric>=2.4.0,<3.0",
    "pydantic~=2.8",
    "fastapi~=0.112",
    "uvicorn[standard]~=0.30",
    "pyyaml~=6.0",
    "redis~=5.0",
    "psycopg2-binary~=2.9",
    "sqlalchemy~=2.0",
    "boto3~=1.34",
    "requests~=2.32",
    "numpy~=1.26",
    "networkx~=3.0",
]

[project.optional-dependencies]
dev = [
    "black~=24.8",
    "ruff~=0.5",
    "pytype~=2024.10.11",
    "pytest~=8.2",
    "pytest-cov~=5.0",
    "httpx~=0.27",
]
docs = [
    "sphinx>=7.0",
    "sphinx-autodoc-typehints>=1.25",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90", "N"]
ignore = ["E501"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytype]
inputs = ["src/esper"]
python_version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src/esper --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/esper"]
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
